pyproject.toml
[build-system]
requires = ["setuptools>=75.8.0", "uv>=0.6.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cs2620-logical-clocks"
version = "0.1.0"
dependencies = [
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
    "tomli>=2.2.0",
    "rich>=13.9.0"
]

[tool.uv]
dev-dependencies = ["ruff>=0.9.0", "mypy>=1.15.0"]
